requirements.txt
Django==1.4.4
South==0.7.6
argparse==1.2.1
django-crispy-forms==1.2.0
gevent==0.13.8
greenlet==0.4.0
gunicorn==0.17.2
pyinotify==0.9.4
python-memcached==1.48
wsgiref==0.1.2
psycopg2==2.4.6
jsonfield==0.9.13
django-mobile==0.2.4
